Refactor: 검색 결과 게시물 리스트 새로운 컴포넌트로 분리
